backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
firebase-admin>=6.5,<7.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
PyJWT>=2.8,<3.0
Pillow>=10.4,<12.0
pytesseract>=0.3.10,<1.0
